+++ src/Order/OrderItem.php
@@ -38,7 +38,7 @@ final class OrderItem
         $this->declinationId = new DeclinationId($data['declinationId']);
         $this->productName = $data['productName'];
         $this->productCode = $data['productCode'];
-        $this->productImageId = $data['productImageId'];
+        $this->productImageId = $data['productImageId'] ?? null;
         $this->price = $data['price'];
         $this->amount = $data['amount'];
         $this->declinationOptions = array_map(static function (array $data) : DeclinationOption {