+++ src/Organisation/OrganisationService.php
@@ -13,6 +13,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Wizaplace\SDK\AbstractService;
 use Wizaplace\SDK\Authentication\AuthenticationRequired;
 use Wizaplace\SDK\Authentication\BadCredentials;
+use Wizaplace\SDK\Order\Order;
 use Wizaplace\SDK\Exception\NotFound;
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
 use Wizaplace\SDK\Exception\UserDoesntBelongToOrganisation;
@@ -379,6 +380,32 @@ class OrganisationService extends AbstractService
     }
 
     /**
+     * Hide a basket, set hide flag to true
+     *
+     * @param string $organisationId
+     * @param string $basketId
+     * @return mixed|null
+     * @throws AuthenticationRequired
+     * @throws NotFound
+     */
+    public function hideBasket(string $organisationId, string $basketId)
+    {
+        $this->client->mustBeAuthenticated();
+
+        try {
+            $responseData = $this->client->post('organisations/'.$organisationId.'/baskets/'.$basketId.'/hide');
+
+            return $responseData;
+        } catch (ClientException $e) {
+            if ($e->getResponse()->getStatusCode() === 404) {
+                throw new NotFound("The organisation doesn't exist", $e);
+            }
+            throw $e;
+        }
+    }
+
+
+    /**
      * Allow to get the organisation's information from a user
      *
      * @param int $userId
@@ -666,6 +693,29 @@ class OrganisationService extends AbstractService
 
                 case Response::HTTP_NOT_FOUND:
                     throw new NotFound("The organisation doesn't exist.", $e);
+
+                default:
+                    throw $e;
+            }
+        }
+    }
+
+    public function getOrder(int $orderId)
+    {
+        $this->client->mustBeAuthenticated();
+
+        try {
+            $response = $this->client->get("organisations/order/{$orderId}");
+
+            return new Order($response);
+        } catch (ClientException $e) {
+            switch ($e->getResponse()->getStatusCode()) {
+                case Response::HTTP_FORBIDDEN:
+                    $response = json_decode($e->getResponse());
+                    throw new \Exception($response->message, Response::HTTP_FORBIDDEN, $e);
+
+                case Response::HTTP_NOT_FOUND:
+                    throw new NotFound("This order doesn't exists.", $e);
 
                 default:
                     throw $e;